refactor: create index.d.ts from other types
