+++ packages/toast-ui.grid/types/editor/index.d.ts
@@ -1,7 +1,7 @@
+import TuiGrid from '../index';
 import { RowKey, CellValue, ListItem } from '../store/data';
 import { ColumnInfo } from '../store/column';
 import { Dictionary } from '../options';
-import { Grid } from '../grid';
 
 export type CheckboxOptions = ListItemOptions & {
   type: 'checkbox' | 'radio';
@@ -10,7 +10,7 @@ export type CheckboxOptions = ListItemOptions & {
 export type PortalEditingKeydown = (ev: KeyboardEvent) => void;
 
 export interface CellEditorProps {
-  grid: Grid;
+  grid: TuiGrid & { usageStatistics: boolean };
   rowKey: RowKey;
   columnInfo: ColumnInfo;
   value: CellValue;
